Modularize ttmetal runtime code
