+++ runtime/lib/ttmetal/CMakeLists.txt
@@ -14,6 +14,7 @@ add_library(TTRuntimeTTMetal
   runtime.cpp
   executor.cpp
   meshshard_utils.cpp
+  kernels.cpp
 )
 # We have to set the C++ standard to 20 because tt-metal requires it
 set_property(TARGET TTRuntimeTTMetal PROPERTY CXX_STANDARD 20)
